pyproject.toml
[tool.poetry]
name = "movie-recommender-system"
version = "1.0.0"
description = "Production-ready movie recommendation system (Python 3.14 compatible)"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.0"
numpy = "^2.0.0"
scikit-learn = "^1.5.0"
mlflow = "^2.16.0"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.9.0"
python-dotenv = "^1.0.0"
requests = "^2.32.0"
sqlalchemy = "^2.0.0"
tqdm = "^4.66.0"
joblib = "^1.4.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.0"
notebook = "^7.2.0"
matplotlib = "^3.9.0"
seaborn = "^0.13.0"
plotly = "^5.24.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.0"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
